requirements.txt
pytorch-lightning==1.5.4
einops==0.3.2
test-tube==0.7.5
kornia==0.4.1
opencv-python==4.5.3.56
matplotlib
jupyter
imageio==2.9.0
imageio-ffmpeg==0.4.2
torch_optimizer
pyquaternion
open3d==0.13.0
numba==0.52.0
black
omegaconf==2.1.1
ipdb

PyMCubes==0.1.2
pyrender==0.1.45
plyfile==0.7.2
black
scipy